Fix good-fences violations by moving CIP code utilities to domain layer
